Support the updated BayIMCO API
The BayIMCO API has changed and now requires the vaccination number
(first, second or boost) to find available appointments. In the case of
a second jab, it also requires the variant used for the first one.

This change updates the checker to support that.

Additionally, filtering for vaccines targeting a specific variant is now
supported.